+++ NEWS.md
@@ -6,9 +6,16 @@
 * Bugfixes
     * Bugfix in `lsm_c_ai()` if only one class and NA values were present
     * Bugfix in `show_correlation()` that first col was lost
+    * Bugfix in `sample_lsm()` and `extract_lsm()` to forward arguments to
+    `calculate_lsm()`
+    * size argument in `sample_lsm()` is now comparable between squares and circles
 * Improvements
     * The values can be added as geom_text in `show_correlation()`
     * `list_lsm()` allows to return all BUT the selected metrics
+    * `sample_lsm()` can now use SpatialPolygons to sample metrics
+    * `sample_lsm()` can now handle SpatialLines to sample metrics
+    * `sample_lsm()` automatically detects provided data type
+    * `extract_lsm()` can now handle SpatialLines to extract metrics
 
 # landscapemetrics 1.0
 * New functions